test: 테스트 파일 패키지구조를 변경하라
